Brace validation fault fix (#111)
