+++ src/events/messageCreate.js
@@ -7,7 +7,7 @@
  * This project uses the AGPLv3 license. Please read the license file before using/adapting any of the code.
  */
 
-const bbEngine = require('../structures/BBTagEngine');
+const bbEngine = require('../structures/bbtag/Engine');
 const Timer = require('../structures/Timer');
 const cleverbotIo = new dep.cleverbotIo({
     user: config.cleverbot.ioid,
